pyproject.toml
[project]
name = "firsttest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "psycopg[binary]>=3.3.2",
    "uvicorn>=0.40.0",
    "python-dotenv>=1.0.0",
    "markdown>=3.10.1",
    "bleach>=6.3.0",
    "pillow>=12.1.0",
    "django-mptt>=1.1",
]
